+++ CHANGELOG.md
@@ -13,6 +13,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ### Changed
 
 - メンバーシップ登録時のポイント特典をChevreで処理するように調整
+- メンバーシップ注文処理をメンバーシップサービスから分離
+- メンバーシップ登録ロックホルダーを注文取引IDに変更
 
 ### Deprecated