+++ CHANGELOG.md
@@ -10,6 +10,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Changed
 
+- update @cinerino/domain
+- メンバーシップ注文配送処理をプロダクト注文配送処理に統合
+
 ### Deprecated
 
 ### Removed